o.k. First try of a automake building enviroment.
Not very comfortably in the moment but it works. For now it can
only create the directory ~/mct-build and build the Crosscompiler.
In ~/mct-build/archives you will find later all archives which downloaded
by buildroot or the Makefile. If you have to rebuild the Crosscompiler
again, the Makefile will use this files, so no new download is nessesary.
So don't delete this files!

More work is nessesary, if i get time i will work on it.
Please test and give some reports back.


git-svn-id: http://www.openmct.org/svn/branches/carsten@345 052c2572-4cfa-0310-8799-9599af202008