Splitting the /etc/rc.d/rcS to second files; at the end of rcS now rcS.second is called, the reason for this the ro attribut of the /etc/rc.d/rcS file, the mini_fo filesystem is just here mounted so the file itself is changable but this takes no effect
The rcS file is staticly and here are no changes nessesary, all changes on the boot have to be done in rcS.second
but remeber: for personaly settings use the /etc/rc.d/rcS.local !!

I added a new target to the main Makefile. With "make rebuild" a new actualy Flashimage will build. This is done without
a rebuild of packages! It's just e rebuild of the newroot directory. This helps me to build shortly new images while working
in /etc.


git-svn-id: http://www.openmct.org/svn/trunk@540 052c2572-4cfa-0310-8799-9599af202008